haciendo cambios de archivos en nextauth
