+++ src/app/api/auth/[...nextauth]/options.js
@@ -1,0 +1,32 @@
+import CredentialsProvider from 'next-auth/providers/credentials';
+
+export const authOptions = {
+    providers: [
+      CredentialsProvider({
+        name: 'Credentials',
+        credentials: {
+          email: { label: 'Email', type: 'text' },
+          password: { label: 'Password', type: 'password' },
+        },
+        async authorize(credentials) {
+          const res = await fetch('http://localhost:3002/api/auth/login', {
+            method: 'POST',
+            headers: { 'Content-Type': 'application/json' },
+            body: JSON.stringify(credentials),
+          });
+  
+          const user = await res.json();
+  
+          if (res.ok && user) {
+            return user;
+          } else {
+            throw new Error(user.error || 'Error al iniciar sesión');
+          }
+        },
+      }),
+    ],
+    pages: {
+      signIn: '/login',
+    },
+    secret: process.env.NEXTAUTH_SECRET,
+}
